+++ Sources/SemVer/Parser/VersionParser+Version.swift
@@ -13,9 +13,9 @@ extension VersionParser {
             throw VersionError.invalidFormat
         }
 
-        let major = Int(result.major)!
-        let minor = Int(result.minor)!
-        let patch = Int(result.patch)!
+        let major = UInt64(result.major)!
+        let minor = UInt64(result.minor)!
+        let patch = UInt64(result.patch)!
 
         let coreVersion = CoreVersion(major: major, minor: minor, patch: patch)
